Move IgnoreLineEndings to org.gradle.work
