+++ subprojects/core-api/src/main/java/org/gradle/work/IgnoreLineEndings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.tasks;
+package org.gradle.work;
 
 import org.gradle.api.Incubating;
